backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
httpx>=0.26.0
PyJWT>=2.8.0
